add pseudocode, move methods to calendar file (with issues)
